♻️ Refactor client components
